move tests to int_udf
